Add boring --plain option to disable emoji encoding
Note that Twitter will strip certain chars from tweets like trailing
linebreaks. In an attempt to overcome this, tweetcat will append one
extra linebreak at the end of all data that is received. This is not
very convenient though, as tweetcat have no idea if the linebreak was
there when the data was sent. But if you keep each message below the
size of one tweet this is usually the decired effect.